requirements.txt
alembic==1.16.5
amqp==5.3.1
annotated-types==0.7.0
anthropic==0.68.0
anyio==4.7.0
bcrypt==4.3.0
billiard==4.2.1
aioboto3==15.5.0
boto3==1.40.61
botocore==1.40.61
cachetools==5.5.2
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==44.0.0
databases==0.9.0
distro==1.9.0
dnspython==2.7.0
docopt==0.6.2
docstring_parser==0.17.0
ecdsa==0.19.0
email_validator==2.2.0
et_xmlfile==2.0.0
fastapi==0.115.6
fastapi-limiter==0.1.6
google-ai-generativelanguage==0.4.0
google-api-core==2.25.1
google-api-python-client==2.182.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.3.2
googleapis-common-protos==1.70.0
greenlet==3.1.1
grpcio==1.75.0
grpcio-status==1.62.3
gunicorn==20.1.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.6.4
httpx[http2]==0.28.1
hyperframe==6.1.0
idna==3.10
jiter==0.10.0
jmespath==1.0.1
kombu==5.4.2
lxml==6.0.1
Mako==1.3.10
MarkupSafe==3.0.2
mysql-connector-python==9.1.0
numpy==2.0.2
openai==1.98.0
openpyxl==3.1.5
orjson==3.11.3
pandas==2.2.3
passlib==1.7.4
pillow==11.3.0
pipreqs==0.4.13
pyotp==2.9.0
qrcode[pil]==8.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
proto-plus==1.26.1
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.10.4
pydantic-settings==2.6.1
pydantic_core==2.27.2
PyMySQL==1.1.1
pyparsing==3.2.4
pypdf==6.0.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.0.1
python-jose[cryptography]==3.5.0
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
razorpay==1.4.2
redis==5.2.1
requests==2.32.3
rsa==4.9
s3transfer==0.14.0
setuptools==79.0.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.2.0
urllib3
uvicorn==0.34.0
vine==5.1.0
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
yarg==0.1.10
reportlab
tenacity
aiomysql
locust
exponent-server-sdk==2.1.0
firebase-admin

psutil>=5.9.0  
sentry-sdk[fastapi]>=1.40.0 
